move do_export() to separate cpp
Attempting to avoid ICE on cuda
