+++ components/scream/src/control/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(SCREAM_CONTROL_SOURCES
   atmosphere_driver.cpp
   surface_coupling.cpp
+  surface_coupling_export.cpp
 )
 
 set(SCREAM_CONTROL_HEADERS
